♻️ refactor: extract coretools sub-package from internal/agent
Move tool implementations (bash, read, write, edit) and output normalizer
into internal/agent/coretools. The builder glue (coretools_builder.go)
stays in agent since it depends on runnerSession. One new export: coretools.New.
