+++ internal/agent/coretools/normalize.go
@@ -1,4 +1,4 @@
-package agent
+package coretools
 
 import (
 	"context"
