+++ internal/agent/coretools/tools_test.go
@@ -1,4 +1,4 @@
-package agent
+package coretools
 
 import (
 	"context"
